pyproject.toml
 [build-system]
 requires = ["setuptools>=69", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "labelling_task"
 version = "0.1.0"
 description = "Multi-tenant data labelling backend (tasks, allocation, review, metrics)."
 readme = "README.md"
 requires-python = ">=3.9,<3.10"
 dependencies = [
   "fastapi>=0.110,<1.0",
   "uvicorn[standard]>=0.27,<1.0",
   "pydantic>=2.5,<3.0",
   "pydantic-settings>=2.1,<3.0",
   "motor>=3.3,<4.0",
   "redis>=5.0,<6.0",
   "python-jose[cryptography]>=3.3,<4.0",
   "boto3>=1.34,<2.0",
   "python-multipart>=0.0.7,<1.0",
 ]
 
 [project.optional-dependencies]
 dev = [
   "pytest>=7.4,<9.0",
   "pytest-asyncio>=0.23,<1.0",
   "httpx>=0.26,<1.0",
   "ruff>=0.3,<1.0",
 ]
 
 [tool.setuptools]
 package-dir = {"" = "src"}
 
 [tool.setuptools.packages.find]
 where = ["src"]
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 testpaths = ["tests"]
 
 [tool.ruff]
 target-version = "py39"
 line-length = 100
